+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+-   `@lumx/react`:
+    -   `TextField`: fix prop type
+    -   `TextField`: fix prop forwarding
+
 ## [4.7.0][] - 2026-03-09
 
 ### Fixed